Add entity_label to alert and rework fallback
Though Slack does appear to do hostname lookups
on the items that appear in the popup
notifications in OS X, we cannot always rely
on the hostname being accurate.

To solve this, adding we will add entity_label
to the content of the alert as this will allow
users to see the friendly "label" of an entity.

Example:
```
...
    "entity": {
        "id": "enphGlgpQy",
        "label": "test",
...
```

This commit also reworks the "fallback" message
that we send to slack.

The fallback message is what Slack uses to
populate the contents of a popup message that
is not rendered directly in a Slack client.

Things like popup notifications on iOS, or
OSX, etc, use this "fallback" to generate
a preview of the alert message.

By adding better content to this message,
admins responding to alerts will be able to
determine if they need to respond to the alert
or not.

Signed-off-by: Evan Gray <evanscottgray@gmail.com>